+++ zubhub_frontend/zubhub/src/views/PageWrapper.jsx
@@ -107,7 +107,7 @@ function PageWrapper(props) {
   const profileMenuOpen = Boolean(anchor_el);
 
   var notif = {
-    type: 1,
+    type: 4,
     id: '26895b6a-2378-464f-bd45-c5f2b3f88229',
     recipient: {
       id: 'b112451f-9c90-437b-96fd-5cabc389435b',
@@ -142,7 +142,7 @@ function PageWrapper(props) {
       members_count: null,
       role: 'creator',
     },
-    link: 'youtube.com',
+    link: 'https://www.youtube.com',
     viewed: false,
     time: '2022-03-12T21:30:57.732Z',
     // time: '2022-03-12T01:53:51.431Z',
